small fix to player.c (#323)

+++ src/melee/pl/player.h
@@ -219,7 +219,7 @@ void Player_SetUnk9C(s32 slot, s32 unk9C);
 HSD_GObj* Player_GetEntity(s32 slot);
 HSD_GObj* Player_GetEntityAtIndex(s32 slot, s32 index);
 void Player_SwapTransformedStates(s32 slot, s32 arg1, s32 arg2);
-s16 Player_GetDamage(s32 slot);
+s32 Player_GetDamage(s32 slot);
 void Player_SetHUDDamage(s32 slot,s32 arg1);
 void Player_SetHPByIndex(s32 slot, s32 arg1, s32 arg2);
 s16 Player_GetOtherStamina(s32 slot);
